Data.Aeson.Validation

Schema validation

data Schema

An opaque JSON Schema.

The <> operator is used to create a sum Schema that, when applied to a Value, first tries the left Schema, then falls back on the right one if the left one fails.

satisfies :: Schema -> Value -> Bool

Does the Value satisfy the Schema?

Boolean schemas

bool :: Schema

Any Bool.

Examples

>>> satisfies bool (Bool True)
True

>>> satisfies bool (Bool False)
True

true :: Schema

Bool True.

Examples

>>> satisfies true (Bool True)
True

>>> satisfies true (Bool False)
False

false :: Schema

Bool False.

Examples

>>> satisfies false (Bool True)
False

>>> satisfies false (Bool False)
True

Number schemas

number :: Schema

Any Number.

Examples

>>> satisfies number (Number 1.0)
True

integer :: Schema

Any integer Number.

Examples

>>> satisfies integer (Number 1.0)
True

>>> satisfies integer (Number 1.5)
False

someNumber :: (Scientific -> Bool) -> Schema

Some Number.

Examples

>>> satisfies (someNumber (> 5)) (Number 6)
True

someInteger :: (Integer -> Bool) -> Schema

Some integer Number.

Examples

>>> satisfies (someInteger (> 5)) (Number 6.0)
True

>>> satisfies (someInteger (> 5)) (Number 6.5)
False

String schemas

string :: Schema

Any String.

Examples

>>> satisfies string (String "foo")
True

theString :: Text -> Schema

An exact String.

Examples

>>> satisfies (theString "foo") (String "foo")
True

>>> satisfies (theString "foo") (String "bar")
False

someString :: (Text -> Bool) -> Schema

Some String.

Examples

>>> satisfies (someString (\s -> Text.length s > 5)) (String "foobar")
True

>>> satisfies (someString (\s -> Text.length s > 5)) (String "foo")
False

regex :: Text -> Schema

A String that matches a regular expression.

Examples

>>> satisfies (regex "a+b") (String "xaaabx")
True

>>> satisfies (regex "c{2}") (String "cd")
False

Object schemas

data Field

An opaque object Field.

Create a Field with .: or .:?, and bundled into a Schema using object or object'

object :: [Field] -> Schema

An Object, possibly with additional unvalidated fields.

To match any Object, use object [].

object' :: [Field] -> Schema

An Object with no additional fields.

The ' mark means "strict" as in foldl', because object' matches Objects more strictly than object.

(.:) :: Text -> Schema -> Field infixr 5

A required Field.

(.:?) :: Text -> Schema -> Field infixr 5

An optional Field.

Array schemas (homogenous lists)

array :: Schema -> Schema

A "homogenous" Array of any size.

The array need not be truly homogenous; it simply has the same Schema applied to each element. However, the Schema could be anything, or composed of many alternatives using <>.

Examples

>>> satisfies (array bool) (Array [Bool True, Bool False])
True

>>> satisfies (array anything) (Array [Bool True, String "foo"])
True

>>> satisfies (array integer) (Array [Number 1.5])
False

sizedArray :: Int -> Int -> Schema -> Schema

A sized (inclusive), "homogenous" (see note above) Array. Use minBound or maxBound for an unbounded edge.

Examples

>>> satisfies (sizedArray 1 2 bool) (Array [Bool True])
True

>>> satisfies (sizedArray 1 2 bool) (Array [Bool True, Bool True, Bool False])
False

Set schemas (homogenous, unique lists)

set :: Schema -> Schema

A "homogenous" (see note above), unique Array of any size.

Examples

>>> satisfies (set bool) (Array [Bool True])
True

>>> satisfies (set bool) (Array [Bool True, Bool True])
False

sizedSet :: Int -> Int -> Schema -> Schema

A sized (inclusive), "homogenous" (see note above), unique Array. Use minBound or maxBound for an unbounded edge.

Examples

>>> satisfies (sizedSet 1 1 string) (Array [String "foo"])
True

>>> satisfies (sizedSet 1 1 string) (Array [String "foo", String "bar"])
False

Tuple schemas (heterogeneous lists)

tuple :: [Schema] -> Schema

A heterogeneous Array exactly as long as the given list of Schemas.

Examples

>>> satisfies (tuple [bool, string]) (Array [Bool True, String "foo"])
True

Miscellaneous schemas

anything :: Schema

Any Value whatsoever, including Null.

Examples

>>> satisfies anything Null
True

>>> satisfies anything (Bool True)
True

nullable :: Schema -> Schema

Modify a Schema to additionally accept Null.

nullable is idempotent:

nullable a = nullable (nullable a)

Examples

>>> satisfies (nullable bool) (Bool True)
True

>>> satisfies (nullable bool) Null
True

inverse :: Schema -> Schema

Succeed whenever the given Schema fails, and vice versa.

inverse . inverse = id

Examples

>>> satisfies (inverse bool) (Bool True)
False

>>> satisfies (inverse bool) (String "foo")
True