Data.Aeson.Validation

Description

JSON schema validation.

Schema validation

data Schema

An opaque JSON Schema.

schema :: Schema -> Value -> Bool

Does the Value satisfy the Schema?

Boolean schemas

bool :: Schema

Any Bool.

Examples

>>> schema bool (Bool True)
True

>>> schema bool (Bool False)
True

true :: Schema

Bool True.

Examples

>>> schema true (Bool True)
True

>>> schema true (Bool False)
False

false :: Schema

Bool False.

Examples

>>> schema false (Bool True)
False

>>> schema false (Bool False)
True

Number schemas

number :: Schema

Any Number.

Examples

>>> schema number (Number 1.0)
True

integer :: Schema

Any integer Number.

Examples

>>> schema integer (Number 1.0)
True

>>> schema integer (Number 1.5)
False

someNumber :: (Scientific -> Bool) -> Schema

Some Number.

Examples

>>> schema (someNumber (> 5)) (Number 6)
True

someInteger :: (Integer -> Bool) -> Schema

Some integer Number.

Examples

>>> schema (someInteger (> 5)) (Number 6.0)
True

>>> schema (someInteger (> 5)) (Number 6.5)
False

String schemas

string :: Schema

Any String.

Examples

>>> schema string (String "foo")
True

theString :: Text -> Schema

An exact String. This is what the OverloadedStrings instance uses when making a Schema from a string literal.

Examples

>>> schema (theString "foo") (String "foo")
True

>>> schema (theString "foo") (String "bar")
False

someString :: (Text -> Bool) -> Schema

Some String.

Examples

>>> schema (someString (\s -> Text.length s > 5)) (String "foobar")
True

>>> schema (someString (\s -> Text.length s > 5)) (String "foo")
False

regex :: Text -> Schema

A String that matches a regular expression.

Examples

>>> schema (regex "a+b") (String "xaaabx")
True

>>> schema (regex "c{2}") (String "cd")
False

Object schemas

data Field

An opaque object Field.

Create a Field with .: or .:?, and bundle it into a Schema using object or object'

object :: [Field] -> Schema

An Object, possibly with additional unvalidated fields.

To match any Object, use object [].

object' :: [Field] -> Schema

An Object with no additional fields.

The ' mark means "strict" as in foldl', because object' matches Objects more strictly than object.

(.:) :: Text -> Schema -> Field infixr 5

A required Field.

(.:?) :: Text -> Schema -> Field infixr 5

An optional Field.

Array schemas (homogenous lists)

array :: Schema -> Schema

A "homogenous" Array of any size.

The array need not be truly homogenous; it simply has the same Schema applied to each element. However, the Schema could be anything, or composed of many alternatives using <>.

Examples

>>> schema (array bool) (Array [Bool True, Bool False])
True

>>> schema (array anything) (Array [Bool True, String "foo"])
True

>>> schema (array integer) (Array [Number 1.5])
False

sizedArray :: Int -> Int -> Schema -> Schema

A sized (inclusive), "homogenous" (see note above) Array. Use minBound or maxBound for an unbounded edge.

Examples

>>> schema (sizedArray 1 2 bool) (Array [Bool True])
True

>>> schema (sizedArray 1 2 bool) (Array [Bool True, Bool True, Bool False])
False

Set schemas (homogenous, unique lists)

set :: Schema -> Schema

A "homogenous" (see note above), unique Array of any size.

Examples

>>> schema (set bool) (Array [Bool True])
True

>>> schema (set bool) (Array [Bool True, Bool True])
False

sizedSet :: Int -> Int -> Schema -> Schema

A sized (inclusive), "homogenous" (see note above), unique Array. Use minBound or maxBound for an unbounded edge.

Examples

>>> schema (sizedSet 1 1 string) (Array [String "foo"])
True

>>> schema (sizedSet 1 1 string) (Array [String "foo", String "bar"])
False

Tuple schemas (heterogeneous lists)

tuple :: [Schema] -> Schema

A heterogeneous Array exactly as long as the given list of Schemas.

Examples

>>> schema (tuple [bool, string]) (Array [Bool True, String "foo"])
True

Miscellaneous schemas

anything :: Schema

Any Value whatsoever, including Null.

Examples

>>> schema anything Null
True

>>> schema anything (Bool True)
True

nullable :: Schema -> Schema

Modify a Schema to additionally accept Null.

nullable is idempotent:

nullable = nullable . nullable

Examples

>>> schema (nullable bool) (Bool True)
True

>>> schema (nullable bool) Null
True